Fix pointer-handling bug in write_random_seed().
Thanks to Joshua Rogers for the report: if writing the random seed
took more than one call to write(2), then we'd add the wrong value to
the data pointer between calls. Happily this is very unlikely to have
actually happened, because random seed files are tiny and surely
write() will in practice manage to write it all first time. But still
needs fixing.